restaurants/requirements.txt
asgiref==3.4.1
Django==3.2.15
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
mock==4.0.3
psycopg2==2.9.3
psycopg2-binary==2.9.3
PyJWT==2.5.0
pytz==2022.2.1
sqlparse==0.4.2
tzdata==2022.2
flake8